+++ test/helpers/test-data-source.ts
@@ -24,7 +24,9 @@ import { CreatePlantingSpots1780000000010 } from '../../src/database/migrations/
 import { AddPlantingSpotIdToPlants1780000000011 } from '../../src/database/migrations/1780000000011-AddPlantingSpotIdToPlants';
 import { AddExpiresAtToQrs1780000000012 } from '../../src/database/migrations/1780000000012-AddExpiresAtToQrs';
 import { CreateSpaceInvitations1780000000013 } from '../../src/database/migrations/1780000000013-CreateSpaceInvitations';
+import { AddAppRoleToAccounts1780000000014 } from '../../src/database/migrations/1780000000014-AddAppRoleToAccounts';
 import { AlterPlantSpeciesEnrich1780000000014 } from '../../src/database/migrations/1780000000014-AlterPlantSpeciesEnrich';
+import { AddTenantRelationshipIndexesAndFks1780000000015 } from '../../src/database/migrations/1780000000015-AddTenantRelationshipIndexesAndFks';
 
 const TEST_ENTITIES = [
   AccountEntity,
@@ -54,7 +56,9 @@ const TEST_MIGRATIONS = [
   AddPlantingSpotIdToPlants1780000000011,
   AddExpiresAtToQrs1780000000012,
   CreateSpaceInvitations1780000000013,
+  AddAppRoleToAccounts1780000000014,
   AlterPlantSpeciesEnrich1780000000014,
+  AddTenantRelationshipIndexesAndFks1780000000015,
 ];
 
 export function getTestDataSourceOptions(): DataSourceOptions {